+++ package.json
@@ -17,8 +17,8 @@
     "prettier": "^3.3.3"
   },
   "dependencies": {
-    "react-calendar": "^5.1.0",
     "bcrypt": "^5.1.1",
-    "jsonwebtoken": "^9.0.2"
+    "jsonwebtoken": "^9.0.2",
+    "react-calendar": "^5.1.0"
   }
 }
