+++ packages/react-frontend/src/components/WorkoutDetail.jsx
@@ -1,10 +1,9 @@
 /* eslint-disable react/prop-types */
-import { useState } from "react";
-import "./workoutdetail.css";
+import styles from "./workoutdetail.module.css";
 
 const WorkoutDetail = ({ key, workout }) => {
   return (
-    <div className="workout">
+    <div className={styles.workout}>
       <h3>Workout Detail</h3>
       <p>Name: {workout.name}</p>
       <p>Type: {workout.workoutType}</p>
