+++ packages/react-frontend/src/components/workoutdetail.module.css
