+++ packages/react-frontend/src/main.jsx
@@ -4,7 +4,7 @@ import ReactDOMClient from "react-dom/client";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from "./home/Home";
 import Calendar from "./calendar/Calendar";
-import DateDetail from "./calendar/DateDetail";
+import DateDetail from "./components/DateDetail";
 import "./styles/global.css";
 
 const container = document.getElementById("root");
